add generator for monitoring endpoints
